git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20181101_acota_q10151@12855 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/preview_preprocessing.php
@@ -90,7 +90,6 @@ if (isset($qlpreview_path) && !in_array($extension, $qlpreview_exclude_extension
     if (file_exists($target)){$newfile = $target;debug("qlpreview success!");}
     }
 
-
 /* ----------------------------------------
     Try InDesign - for CS5 (page previews)
    ----------------------------------------
@@ -106,9 +105,9 @@ if ($exiftool_fullpath!=false)
             
             $n=0;
             foreach ($indd_thumbs as $indd_page){
-                // echo $indd_page;
-                $pagescommand.=" ".$target."_".$n;
-                base64_to_jpeg( str_replace("base64:","",$indd_page), $target."_".$n);
+                $target_pg = str_replace(".jpg","_" . $n . ".jpg", $target);
+                $pagescommand.=" " . $target_pg;
+                base64_to_jpeg( str_replace("base64:","",$indd_page), $target_pg);
                 
                 $n++;
             }
@@ -118,12 +117,15 @@ if ($exiftool_fullpath!=false)
         // process jpgs as a pdf so the existing pdf paging code can be used.   
         if (is_array($indd_thumbs)){
             $file=get_resource_path($ref,true,"",false,"pdf");      
-            $jpg2pdfcommand = $convert_fullpath . " ".$pagescommand." " . $file;
+            $jpg2pdfcommand = $convert_fullpath . " " . $pagescommand . " " . escapeshellarg($file);
+
             $output=run_command($jpg2pdfcommand);
+
             $n=0;
             foreach ($indd_thumbs as $indd_page){
-                if (file_exists($target."_".$n)){   
-                    unlink($target."_".$n);
+                $target_pg = str_replace(".jpg","_" . $n . ".jpg", $target);
+                if (file_exists($target_pg)){   
+                    unlink($target_pg);
                 }
                 $n++;
             }
@@ -252,7 +254,11 @@ if (($extension=="cr2" || $extension=="nef" || $extension=="dng" || $extension==
             // check for nef -otherimage failure
             if ($extension=="nef"&&!filesize_unlimited($target)>0)
                 {
-                unlink($target);    
+                if(file_exists($target))
+                    {
+                    unlink($target);
+                    }
+
                 $bin_tag=" -previewimage ";
                 //2nd attempt
                 $cmd=$exiftool_fullpath.' -b '.$bin_tag.' '.escapeshellarg($file).' -w %d%f.jpg';
@@ -318,7 +324,7 @@ if ( (($extension=="pages") || ($extension=="numbers") || (!isset($unoconv_path)
    ----------------------------------------
 */
 global $unoconv_extensions;
-if (in_array($extension,$unoconv_extensions) && isset($unoconv_path) && !isset($newfile))
+if (in_array($extension,$unoconv_extensions) && $extension!='pdf' && isset($unoconv_path) && !isset($newfile))
     {
     global $config_windows;
     $unocommand=$unoconv_path . "/unoconv";
@@ -389,7 +395,7 @@ if (in_array($extension,$calibre_extensions) && isset($calibre_path) && !isset($
     $basename_minus_extension=remove_extension($path_parts['basename']);
     $pdffile=$path_parts['dirname']."/".$basename_minus_extension.".pdf";
 
-    $cmd="xvfb-run ". $calibrecommand . " " . escapeshellarg($file) . " " .$pdffile." ";
+    $cmd="xvfb-run ". $calibrecommand . " " . escapeshellarg($file) . " " . escapeshellarg($pdffile) ." ";
     $wait=run_command($cmd);
 
     if (file_exists($pdffile))
@@ -983,7 +989,24 @@ if ((!isset($newfile)) && (!in_array($extension, $ffmpeg_audio_extensions))&& (!
 # If a file has been created, generate previews just as if a JPG was uploaded.
 if (isset($newfile))
     {
-    create_previews($ref,false,"jpg",$previewonly,false,$alternative);  
-    }
+    if($GLOBALS['non_image_types_generate_preview_only'] && in_array($extension, $GLOBALS['non_image_types']))
+        {
+        $file_used_for_previewonly = get_resource_path($ref, true, "tmp", false, "jpg");
 
-?>
+        if(copy($newfile, $file_used_for_previewonly))
+            {
+            $previewonly = true;
+            debug("preview_preprocessing: changing previewonly = true for non-image file");
+            }
+        }
+
+    create_previews($ref,false,"jpg",$previewonly,false,$alternative);
+
+    if(
+        $GLOBALS['non_image_types_generate_preview_only']
+        && in_array($extension, $GLOBALS['non_image_types'])
+        && file_exists($file_used_for_previewonly))
+        {
+        unlink($file_used_for_previewonly);
+        }
+    }
